Remove `code` and `scope`, replace with `suite`
The Zuse IR defined a `scope` and `code` statement that
 had the same syntax, but different semantics. `code` would
theoretically not create a new scope, but `scope` would.
Since we decided to change to C-style scoping, these are
redundant, so combining them into one is wise. Additionally,
I changed it to `suite`, to differentiate between the Zuse IR
set of statements/expressions and those which are just keys
in the Zuse DSL.