Handle division by zero and type coercion
Refactored code so that it’s data-driven instead of a bunch
of if/else statements. Might be wise to refactor all the code
to behave this way. Also added type coercion for all
expressions. I may be missing some. Divide by 0 now
divides by 1. Not sure the best way to handle that.